+++ src/test/java/pojos/TagsPojo.java
@@ -11,10 +11,10 @@ public class TagsPojo {
     }
   ]
      */
-    private Integer id;
+    private Long id;
     private String name;
 
-    public TagsPojo(Integer id, String name) {
+    public TagsPojo(Long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -22,11 +22,11 @@ public class TagsPojo {
     public TagsPojo() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
